+++ src/main/java/ru/hh/resumebuilderbot/database/dao/UserDAOImpl.java
@@ -2,18 +2,13 @@ package ru.hh.resumebuilderbot.database.dao;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
-import org.hibernate.Criteria;
 import org.hibernate.SessionFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.hh.resumebuilderbot.database.dao.base.GenericDAOImpl;
 import ru.hh.resumebuilderbot.database.model.User;
-
 import javax.persistence.NoResultException;
 import javax.persistence.TypedQuery;
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
 import javax.validation.constraints.NotNull;
 
 @Singleton
@@ -27,7 +22,7 @@ public class UserDAOImpl extends GenericDAOImpl<User, Integer> implements UserDA
 
     private final static String USER_BY_TELEGRAM_ID = "FROM TelegramUser WHERE telegram_id = :telegram_id";
 
-    public User getByTelegramId(Integer telegramId){
+    public User getByTelegramId(long telegramId){
         TypedQuery<User> query = getCurrentSession().createQuery(USER_BY_TELEGRAM_ID, User.class);
         query.setParameter("telegram_id", telegramId);
         User user = null;
